convert to prosemirror node
